Adds Mentor model with related relations
