+++ app/Mentor.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Mentor extends Model
+{
+    public function user(){
+        return $this->belongsTo(User::class);
+    }
+
+    public function sessions(){
+        return $this->hasMany(Session::class);
+    }
+
+    public function students(){
+        return $this->hasMany(Student::class);
+    }
+}
