+++ app/Session.php
@@ -6,8 +6,8 @@ use Illuminate\Database\Eloquent\Model;
 
 class Session extends Model
 {
-    public function user(){
-        return $this->belongsTo(User::class);
+    public function mentor(){
+        return $this->belongsTo(Mentor::class);
     }
 
     public function student(){
